Update notes.md
Added notes for JavaScript and the Midterm review.

Note: Before taking the midterm, go back and look through all these notes and the midterm topics. If there is anything from the midterm topics that is not in the notes.md, look up what they are and add them later. Have added placeholders for such topics.